(PC-7905): Moving a use case into the core api folder
